src/requirements.txt
python-telegram-bot[all]==20.0
